通知代码下沉到 manager 层方便复用
